backend-monolith/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
aiokafka==0.11.0
httpx==0.27.0
python-multipart==0.0.9
python-socketio[asyncio_client]==5.11.3
elasticsearch[async]==8.13.0
aiosmtplib==3.0.1
